522: v1.6: add recheck and modify tasks
