+++ app/server.py
@@ -39,6 +39,7 @@ from routes.login import login
 from routes.user import user_blueprint
 from routes.tasks import tasks
 from routes.upload import upload
+from routes.recheck import recheck
 
 from server_consts import UPLOAD_FOLDER, ALLOWED_EXTENSIONS, DOCUMENT_TYPES, TABLE_COLUMNS, URL_DOMEN
 
@@ -61,6 +62,7 @@ app.register_blueprint(login, url_prefix='/login')
 app.register_blueprint(user_blueprint, url_prefix='/user')
 app.register_blueprint(tasks, url_prefix='/tasks')
 app.register_blueprint(upload, url_prefix='/upload')
+app.register_blueprint(recheck, url_prefix='/recheck')
 
 app.logger.addHandler(get_logging_stdout_handler())
 app.logger.propagate = False
@@ -87,45 +89,6 @@ def signup():
 
 # Main chapters req handlers:
 
-
-
-@app.route("/recheck/<check_id>", methods=["GET"])
-@login_required
-def recheck(check_id):
-    if not current_user.is_admin:
-        abort(403)
-    oid = ObjectId(check_id)
-    check = db_methods.get_check(oid)
-
-    if not check:
-        abort(404)
-    
-    # write files (original and pdf) to filestorage
-    filepath = join(UPLOAD_FOLDER, f"{check_id}.{check.filename.rsplit('.', 1)[-1]}")
-    pdf_filepath = join(UPLOAD_FOLDER, f"{check_id}.pdf")
-    db_methods.write_file_from_db_file(oid, filepath)
-    db_methods.write_file_from_db_file(ObjectId(check.conv_pdf_fs_id), pdf_filepath)
-    
-    check.is_ended = False
-    db_methods.update_check(check)
-    task = create_task.delay(check.pack(to_str=True))  # add check to queue
-    db_methods.add_celery_task(task.id, check_id)  # mapping celery_task to check (check_id = file_id)
-    if request.args.get('api'):
-        return {'task_id': task.id, 'check_id': check_id}
-    else:
-        return redirect(url_for('results', _id=check_id))
-
-
-@app.route("/tasks/<task_id>", methods=["GET"])
-@login_required
-def get_status(task_id):
-    task_result = AsyncResult(task_id)
-    result = {
-        "task_id": task_id,
-        "task_status": task_result.status,
-        "task_result": task_result.result,
-    }
-    return jsonify(result), 200
 
 @app.route("/results/<string:_id>", methods=["GET"])
 def results(_id):
